Updates to remove CODEC and add utility methods to IamResourceType enum; IamResourceTypeEnum now uses SPEND_PROFILE
