+++ src/test/java/bio/terra/service/auth/iam/IamResourceTypeTest.java
@@ -1,34 +1,31 @@
-package bio.terra.common;
+package bio.terra.service.auth.iam;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 
 import bio.terra.common.category.Unit;
 import bio.terra.model.IamResourceTypeEnum;
-import bio.terra.service.auth.iam.IamResourceType;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.springframework.test.context.ActiveProfiles;
 
 @ActiveProfiles({"google", "unittest"})
 @Category(Unit.class)
-public class IamResourceTypeCODECTest {
+public class IamResourceTypeTest {
 
   @Test
   public void testAllEnumValues() {
     for (IamResourceType resource : IamResourceType.values()) {
       assertThat(
-          "CODEC encodes and decodes",
-          IamResourceTypeCODEC.toIamResourceType(
-              IamResourceTypeCODEC.toIamResourceTypeEnum(resource).toString()),
+          "ENUM encodes and decodes",
+          IamResourceType.toIamResourceType(IamResourceType.toIamResourceTypeEnum(resource)),
           equalTo(resource));
     }
 
     for (IamResourceTypeEnum resource : IamResourceTypeEnum.values()) {
       assertThat(
-          "CODEC encodes and decodes",
-          IamResourceTypeCODEC.toIamResourceTypeEnum(
-              IamResourceTypeCODEC.toIamResourceType(resource.toString())),
+          "ENUM encodes and decodes",
+          IamResourceType.toIamResourceTypeEnum(IamResourceType.toIamResourceType(resource)),
           equalTo(resource));
     }
   }
